third_party/blink/web_tests/wpt_internal/view-transition-on-navigation/at-rule-opt-out-via-cssom.html

<!DOCTYPE html>
<title>View Transitions: @view-transition opt-out programmatically.</title>
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-2/">
<link rel="author" href="mailto:bokan@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
@view-transition {
    navigation: auto;
}
</style>
<script>
const params = new URLSearchParams(location.search);
const is_new_page = params.has('new');

if (!is_new_page) {
  onload = () => requestAnimationFrame(() => requestAnimationFrame(() => {
      document.styleSheets[0].rules[0].navigation = 'none';
      location.replace(location.href + '?new');
  }));
} else {
  promise_test(() => {
    return new Promise((resolve) => {
        addEventListener('pagereveal', resolve);
    }).then(event => {
        assert_equals(event.viewTransition, null,
            "ViewTransition must not be triggered.");
    });
  });
}
</script>
